backend/requirements.txt
fastapi
uvicorn
alembic
tenacity
psycopg2-binary
python-multipart
pyjwt
pyotp
bcrypt
python-dotenv
requests
celery[redis]